backend/requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.16,<4.0
django-cors-headers>=4.8,<5.0
psycopg[binary]>=3.3,<4.0
spiffworkflow>=3.0,<4.0
RestrictedPython>=7.0,<8.0
